src/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.2.0

# Code Quality
flake8==6.1.0
black==23.9.1
isort==5.12.0
mypy==1.5.1
pre-commit==3.5.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0